+++ api/src/main/java/io/strimzi/api/kafka/model/AutoRestart.java
@@ -5,6 +5,7 @@
 package io.strimzi.api.kafka.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import io.strimzi.crdgenerator.annotations.Description;
 import io.strimzi.crdgenerator.annotations.DescriptionFile;
 import io.sundr.builder.annotations.Buildable;
@@ -23,11 +24,15 @@ import java.util.Map;
     builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
 @JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({"enabled", "maxRestarts"})
 @EqualsAndHashCode
 public class AutoRestart implements UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
+
     private boolean enabled = true;
-    private Map<String, Object> additionalProperties = new HashMap<>(0);
+    private Integer maxRestarts;
+
+    private final Map<String, Object> additionalProperties = new HashMap<>(0);
 
     @Description("Whether automatic restart for failed connectors and tasks should be enabled or disabled")
     public boolean isEnabled() {
@@ -36,6 +41,17 @@ public class AutoRestart implements UnknownPropertyPreserving, Serializable {
 
     public void setEnabled(boolean enabled) {
         this.enabled = enabled;
+    }
+
+    @Description("The maximum number of connector restarts that the operator will try. " +
+            "If the connector remains in a failed state after reaching this limit, it must be restarted manually by the user. " +
+            "Defaults to an unlimited number of restarts.")
+    public Integer getMaxRestarts() {
+        return maxRestarts;
+    }
+
+    public void setMaxRestarts(Integer maxRestarts) {
+        this.maxRestarts = maxRestarts;
     }
 
     @Override